Add patches for different versions of Test::Unit.
Because Mocha effectively replaces the Test::Unit::TestCase#run
method, we need to ensure that any changes to the Test::Unit version
of this method are reflected in the Mocha patches. This has been
neglected since version 2.0.2.

Also added warnings to aid debugging. These currently use
Kernel#warn and therefore go to stderr. We may need to change this
to make it easier to disable e.g. using the existing Mocha Logger.

Also add a warning if an "unverified" version of Test::Unit is
being used. And a warning if no suitable patch can be identified.