Improve documentation for Expectation#raises.
I've had a number of people suggesting an extension to the API
to cope with custom exceptions that have extra constructor
parameters. However, since Mocha just passes the arguments on
to Kernel#raise, it's possible to pass in an instance of an
exception. Thus no change to the API is required, but it does
seem worthwhile pointing this out in the docs.