Fix exception when running rake without test-unit gem.
When test-unit gem >=v2.0.0 was installed but the "use_test_unit_gem" MOCHA_OPTION was not specified, a "comparison of Fixnum with Hash failed" exception was being raised when running the performance tests. This was because bits of the test-unit gem were being loaded accidentally and a Hash was being incorrectly supplied to the TestRunner.run method.