Improved test coverage around stubbing of class methods defined in different ways.
In particular, check that the hierarchy of methods defined on a class and its ancestors is unchanged after a test.

Note this adds loads of duplication and what is probably a very inefficient new Introspection::ClassSnapshot class. I'm sure it's probably possible to do simplify all this with some elegant recursion and probably combine the InstanceSnapshot and ClassSnapshot, but it's Friday afternoon and my brain is too frazzled.