Do not monkey-patch Test::Unit or MiniTest unless we *know* it's ok.
I had assumed both libraries were changing relatively slowly and that
the Mocha Travis CI build in conjunction with the
Gemfile.minitest.latest and Gemfile.test-unit.latest gemfiles would
catch any problems, but this seems to have been overly optimistic.

Also the logic in this set of if statements was flawed - the else
statement was unlikely to ever be reached.