Rename Integration module to MonkeyPatching, deprecating Integration.
This was first done here [1], but then reverted [2] when it was
discovered that Rails/ActiveSupport is using a class within the
Integration module.

This time, we're keeping the Integration module available, but
deprecating it. The only simple alternative to using it at the moment is
to use the MonkeyPatching module, but we hope to release a version of
Mocha soon which will avoid the need to use it at all. At that point, we
can deprecate the use of MonkeyPatching as well.

[1] 2334b922881125e5cd1a0ef77b25c81a13e91269
[2] 1017ea89df6b4b8abe068154be71551a1caf053d