Use new Test::Unit adapter for v2.3.0-2.5.1 instead of monkey patching.
c.f. f4ad99ca8ef44f8f911fc747ce8bb42b23f0746a for the MiniTest
equivalent.

I'm not sure whether or not it was causing a problem, but I decided to
keep things consistent with the MiniTest equivalent by also moving the
version checking code out of the adapter file.

Note that I had to change the hash syntax in the Test::Unit adapter to
work in Ruby 1.8.

This is also the change that demonstrated the problem that I fixed in
e9d05cc2acf3a0bf5fbc78737d7dfa684e12807e.