Make test library integration more explicit.
Based on a suggestion from @lazyatom. Thanks!

Now integration only happens when the newly introduced `Mocha.activate`
method is invoked; previously this happened when certain files were
required.

A side benefit of this is that a bunch of code that was in the top-level
namespace has now been extracted into `Mocha` modules.