Remove troublesome require statements.
These require statements meant that Test::Unit was always loaded when
Mocha was loaded. This is not what we want. It was an oversight when the
code was change to always load all the integration modules.

The checks in Mocha::Integration will ensure that Test::Unit::TestCase
is available if any of these modules is actually included.