Officially support minitest v3.4.0 & test-unit v2.5.2.
Still monkey-patching, although there have been no changes to the
monkey-patched code, so this is just a change in the allowed versions.

Note that I'm planning on releasing this from a non-master branch,
because there's significant work-in-progress on a more permanent
solution that's already in master, but not quite ready for release.