Officially support minitest v3.5.0.
Still monkey-patching, although there have been no changes to the
monkey-patched code, so this is just a change in the allowed
versions.

It looks like there was an oversight in the last release where version
3.3.0 of minitest was left out in the cold. I have rectified this in
this commit.

Note that I'm planning on releasing this from a non-master branch,
because there's significant work-in-progress on a more permanent
solution that's already in master, but not quite ready for release.