Add Google Analytics javascript to Mocha docs.
Customising YARD templates is a bit obscure and this was the best I
could come up with. The javascript appears as the last section within
the content div, but before the footer div. Not ideal, but I think it
should be ok.

The tracking javascript is only displayed if the
`GOOGLE_ANALYTICS_WEB_PROPERTY_ID` environment variable is set.