Officially support test-unit v2.5.3.
Still monkey-patching, although there have been no changes to the
monkey-patched code, so this is just a change in the allowed versions.

Note that I'm planning on releasing this as a 0.12.x patch release so
people with earlier versions of ActiveSupport can use a fairly recent
version of Mocha with the latest version of Test::Unit. At the moment,
earlier versions (e.g. 3.0.x, 3.1.x, 3.2.x) of ActiveSupport do not
work with 0.13.x versions of Mocha.