Officially support minitest v4.4.0
Still monkey-patching.

Note that I'm planning on releasing this as a 0.12.x patch release so
people with earlier versions of ActiveSupport can use a fairly recent
version of Mocha with the latest version of MiniTest. At the moment,
earlier versions (e.g. 3.0.x, 3.1.x, 3.2.x) of ActiveSupport do not
work with 0.13.x versions of Mocha.