Adapt tests to changes in MiniTest output.
The Mocha integration tests are unfortunately dependent on the
exact format of the test results output by MiniTest. This means that
since the format changed in MiniTest v4.6.1, we have to change the
regular expressions used to parse the output.

It would be nicer to use named capture groups for the regular
expression, but these only appear to be available in Ruby >= v1.9.