Revert "Consider stubs on superclasses if none exist on receiver. Fixes #145."
This reverts commit b41254ac391ce2274b89297b6bbb89b179a7df45.

As pointed out by @ccutrer this solution doesn't work in the case where
a method is stubbed on a child class and then another method is stubbed
on the parent class. This is because as soon as the `@mocha` instance
variable is set for the child class, it never tries to use the `@mocha`
instance variable for the parent class. Time for a rethink and another
look at #145.