Fix Mocha::Integration::MiniTest.activate return value.
It looks like I got in a muddle with the levels of indentation.

With this change `.activate` will now return `true` if an integration module
has already been applied. This is a change in the logic, but I think it makes
sense i.e. the return value indicates whether or not Mocha has been integrated
with MiniTest.