Force use of Test::Unit runner, etc in relevant integration tests
Prior to this, I don't think we were really testing the Mocha integration with
Test::Unit much, because, although `TestUnitTest` was a subclass of
`Test::Unit::TestCase`, the important test case instances are the temporary
ones built by `TestRunner#run_as_test` et al. Prior to this change, these would
only have used Test::Unit where MiniTest was not available *at all* i.e. only
in early versions of Ruby and when the MiniTest gem was not loaded.

I have intentionally not fixed the indentation in `test/test_runner.rb` so that
the diff is clearer. I intend to fix the indentation in a separate commit.