Remove unnecessary code in integration tests.
`test/test_helper.rb` already loads MiniTest or Test::Unit, so no need to do
it again here.

Also the "outer" test case just needs to be of a type that can run Mocha
acceptance tests i.e. it's the "inner" test case that needs to be either
a `MiniTest::Unit::TestCase`, `Minitest::Test` or `Test::Unit::TestCase`
as appropriate. The simplest thing to do is to make use of `Mocha::TestCase`
which is already begin defined in `test/test_helper.rb`.