Assume 'mocha' has been required when requiring 'mocha/setup'
So no need to require 'mocha/version' in both files.

Note that 'mocha' is likely to have been auto-required if you are using it
within a bundler `Gemfile`, although be careful to remove any `require: false`
option that may have been on the gem declaration for earlier versions of Mocha.