Allow stubbing method to which any instance responds.
Fixes #200.

Previously we were not calling `Object#respond_to?` on an instance of the
class. By using `Object#allocate` we can instantiate an example instance of
the class so we can interrogate it using `Object#respond_to?`.

Obviously this is not perfect, given that a method might be defined in the time
between stubbing and invocation, but it means that both
`Mocha::ObjectMethods#method_exists?` and `Mocha:ClassMethods#method_exists?`
now work in a similar way.