Require at least Ruby v1.8.7
This is the earliest version of Ruby available on Travis CI and
based on some initial investigation, it isn't currently straightforward
to install earlier versions locally. Thus although at one point Mocha
definitely worked on Ruby v1.8.6, that cannot now be verified.

This change also means that a sensible value will be displayed against
the "Required Ruby Version" on the Rubygems page [1]. At the moment
this is showing as ">= 0".

[1]: https://rubygems.org/gems/mocha