As of Ruby v2.2, there are no test libs in the standard library.
Both Test::Unit and MiniTest were removed [1] from the standard library in
Ruby v2.2.

Since Mocha relies on the presence of either Test::Unit or MiniTest to run its
tests, this commit makes MiniTest an explicit development dependency for
versions of Ruby which do not make it available by default.

[1]: https://bugs.ruby-lang.org/issues/9711#note-12