As of Ruby v2.2, nil is frozen so these tests are irrelevant.
In all these scenarios in the latest versions of Ruby, a `Mocha::StubbingError`
will be raised with the message: "can't stub method on frozen object: nil".

This behaviour is already tested in the more generic `StubbingFrozenObjectTest`
and so it's safe to ignore these tests for the relevant versions of Ruby.