Constrain version of MiniTest in gem spec to v4.x.
Some major changes happened between MiniTest v4 and v5 and I was seeing
exceptions like this one when running the performance tests:

    Benchmarking with 1000 iterations...
    rake aborted!
    NoMethodError: undefined method `output=' for Minitest::Unit:Class
    /Users/jamesmead/Code/freerange/mocha/Rakefile:94:in `benchmark_test_case'
    /Users/jamesmead/Code/freerange/mocha/Rakefile:84:in `block (3 levels) in <top (required)>'
    /Users/jamesmead/Code/freerange/mocha/Rakefile:83:in `each'
    /Users/jamesmead/Code/freerange/mocha/Rakefile:83:in `block (2 levels) in <top (required)>'
    Tasks: TOP => default => test:performance
    (See full trace by running task with --trace)

This is because we're relying on bits of MiniTest which no longer exist in the
same form.

Rather than addressing these problems now, I'm adding a constraint to the gem
spec file in an effort to get the Travis CI build passing under Ruby v2.2 as
soon as possible. It's possible that we'll run into issues with the
MiniTest-specific Gemfiles, but let's see!