Remove MiniTest version constraint from gem spec.
This is a bodgy way to make the performance tests work a bit (after a fashion)
with MiniTest v5. It's not ideal, because there's no longer a global way to
set the output/reporter nor a way to disable MiniTest's autorun mechanism, so
we see some unhelpful output about running 0 tests at the end.

However, I haven't really been keeping an eye on this benchmark for some time
and it's not currently part of the Travis CI build, so it doesn't seem hugely
important. It seems *more* important to remove the gem spec constraint so that
Travis CI builds will run against the latest versions of MiniTest.