Explicitly support matching `Hash` parameters with `includes`.
The normal `Hash#includes?` method is a synonym for `has_key?`; this
would've been the existing matcher behaviour prior to these changes.
The new behaviour is that matchers can also now be given, and for
`Hash` instances they will be matched against the keys of the hash.

The `String` class in Ruby 1.8.7 responds to `all?`, yielding the
whole string, but we actually want all `String` instances to use the
basic `parameter.include?` code path.