Remove version restriction on Redcarpet.
I was seeing the following exception when trying to generate the docs:

    NoMethodError: undefined method `lax_spacing=' for #<RedcarpetCompat:0x007f9a643a85f8>

As of Yard v0.7.4 [1], we no longer need the restriction on the version of
Redcarpet which was introduced in this commit [2] and this restriction seems
to be what's causing the exception mentioned above.

It's possible that we could now remove the dependency on Redcarpet altogether,
because I think Yard now had a built-in markdown processor. However, I haven't
managed to convince myself that such a removal doesn't break some formatting
somewhere, so this seems like the best option for now.

[1]: https://github.com/lsegal/yard/blob/master/CHANGELOG.md#074---december-2-2011
[2]: 4dae54c39001c0ed4d8fd9dd504f7f0204da51cd