Invert sense of environment variable for generating docs.
I was seeing an error in Travis CI builds for Ruby versions earlier than 1.9.2:

    An error occurred while installing redcarpet (3.2.2), and Bundler cannot continue.

This was happening because I'd forgotten to include the `MOCHA_NO_DOCS`
environment variable in the more specific builds (i.e. in `matrix[include]`) in
`.travis.yml`.

Instead of adding this environment variable to all the builds, I've elected to
invert the logic.