Simplify implementation of String#mocha_inspect.
This is an alternative attempt at fixing the problem identified
and addressed in #215 regarding the "pretty printing" of single- and
double-quotes.

In this attempt, I asked myself why we *ever* needed to do the `gsub`
replacement of escaped double-quote characters with single-quote characters.
I couldn't come up with a reason other than to make the tests more readable.

Thus in this commit I've removed the custom implementation of
`String#mocha_inspect` so it falls back to the default `String#inspect`
implementation. As far as I can see the error messages are still sensibly
formatted and by using the `%{}` string literal construction, I think the
tests are still perfectly readable.

I hope that this is a simpler and more robust way to solve the problems
that @neonichu was seeing.