Pass result of yield to caller
When using #yields or #multiple_yields, the expectation will only return a value if it is explicitly specified with a #returns call.
The default behavior of Ruby blocks is to pass the result of the block back to the caller.

This mimics that behavior with expectations that yield. The result of the yielded block is the return value of the expectation unless a return value was explicitly specified. This allows stubbing out the logic of a method that ultimately yields to a given block without breaking its caller's expectation of the return value it receives, vs. a #yield stub implicitly stubbing out the behavior of everything it yields to as well.

In the case of #multiple_yields, the result from the final yield is returned, as if the stubbed method was executing:
yield(first_parameter_group)
yield(second_parameter_group)
…
yield(final_parameter_group)

Fixes https://floehopper.lighthouseapp.com/projects/22289/tickets/14-8687-blocks-return-value-is-dropped-on-stubbed-yielding-methods