Check all accessible methods when asserting visibility
We're planning to use the prepended module stubbing mechanism more widely for
Ruby v2+. This means that the assertion needs to consider methods which are not
defined on `object`. We're happy to do this since all the places where this
assertion is used are looking for a method with an unusual name, i.e. there's no
real possibility that the check would accidentally find the method on another of
the ancestors of `object`.