Always use prepended module stubbing mechanism for Ruby v2+
Previously we only used this mechanism when there were any prepended modules.

Having just chatted about this, we think it would've made more sense if it had
only been used when the stubbed method was defined on a prepended module.

This commit means that we now use the mechanism for Ruby v2 onwards which is
when `Module#prepend` was first introduced.

The mechanism prepends a module to the stubbee and defines the stub method on
that module. This means that calls to the stubbed methods are intercepted by the
implementation on the prepended module and there is no need to remove the
original implementation. This in turn means that it is easier to reinstate the
original behaviour on teardown, i.e. by simply removing the method on the
prepended module to reveal the original method.