Add test for stubbing class method defined on module & aliased
This test was derived from the tests in #244 and was failing before we
moved to always (in Ruby v2+) using the prepended module stubbing
mechanism. We'll probably want to expand it to cover some more
scenarios, but it's useful as it stands.

This test fails in Ruby 1.8.x and we can't find an obvious fix.

The test passes in Ruby 1.9.x due to a change in behaviour of `#owner`
when a class is `extend`ed by a module.

The test passes in Rubies > 2.0 because we now always prepend a module
to avoid overwriting the original method at all.