Only run test in Ruby 1.9 and above
This test fails in Ruby 1.8.x and we can't find an obvious fix.

The test passes in Ruby 1.9.x due to a change in behaviour of
`#owner` when a class is `extend`ed by a module.

The test passes in Rubies > 2.0 because we now always prepend a module
to avoid overwriting the original method at all.

This diff is best viewed using the `ignore-all-space` option.