add tests for method visibility
Add assertions to ensure that we're correctly setting the visibility of
stubbed methods when those methods are defined on a module that's
included in our class.

We're adding these assertions to help ensure we don't add regressions in
future. The behaviour being tested here was incorrect prior to commit
e87c03b068efc48267fbcd5a295514077c52b901 in that the assertions for
`:private` and `:protected` method visibility were failing.