Simplify logic in AnyInstanceMethod#hide_original_method
This is the same change that was made to `ClassMethod` in
e87c03b068efc48267fbcd5a295514077c52b901.

We're now considering methods in all ancestors when determining the
visibility of the original method. This is so that we correctly set the
visibility irrespective of whether or not the method to stub is defined
on the stubbee. In particular, this means that we're now correctly
setting the visibility of stubbed methods when the original was defined
on a superclass.  This change is described by the
`assert_method_visibility` assertions in
test_should_be_able_to_stub_a_protected_superclass_method and
test_should_be_able_to_stub_a_private_superclass_method.

I've had to update four tests around ActiveRecord association behaviour
in Ruby 1.8 and 1.9. Now that we're setting visibility correctly, we
have to use `send` to call the stubbed protected and private methods.