Be more specific in separating Minitest errors & failures
Minitest::UnexpectedError is a sub-class of Minitest:Assertion and so
previously all the errors were showing up in the failures as well.

While this wasn't causing any problems, this was only because none of
the tests relied on this behaviour being correct.

I've checked that Object#instance_of? has existed since at least Ruby
v1.8.7.

My motivation for fixing this problem was #250 and this change should
help me fix that problem in a subsequent commit.