Fix failing test in Ruby 2.3
Ruby bug 12832 is only triggered when we try to get hold of the method
being stubbed. When discussing this a couple of days ago we realised
that we should be able to avoid the problem because we don't need a
reference to the original method when we're using prepended modules
(i.e. in Ruby version 2 and above).

The change in this commit fixes the failing test in Issue272Test. All
other tests continue to pass under Ruby 2.3.

[1]: https://bugs.ruby-lang.org/issues/12832