Fix failures in class_method_test under 1.8.7/1.9.3
I was seeing the following 3 failures when running the tests in this
test case:

1. test_should_hide_original_method(ClassMethodTest) [test/unit/class_method_test.rb:18]:
   <false> expected but was
   <true>.

2. test_should_restore_original_method(ClassMethodTest):
   NameError: undefined method `method_x' for class `Class'

3. test_should_restore_original_method_accepting_a_block_parameter(ClassMethodTest):
   NameError: undefined method `method_x' for class `Class'

Each of these failures was caused by the absence of the `#_method`
method. Invoking this method in `#hide_original_method` was raising an
exception that was being caught by the `rescue NameError` block. I've
used `alias_method` to define the `#_method` method as this matches the
real behaviour in `Mocha::ObjectMethods`.

These tests would pass when run using `rake test` because
`Mocha::ObjectMethods` was being included elsewhere which meant that the
real implementation of `_method` was being invoked.

NOTE. These tests were previously failing in Ruby 1.8.7 and 1.9.3.
Versions 2.0.0 and above use the prepended modules approach which don't
involve the use of `_method`.