No need to set visibility if stubbed method not restored
Fixes #256.

There's no point in considering setting the visibility of the restored
method if we haven't actually restored (i.e. re-defined) it.

It may be possible to further simplify the logic in ClassMethod, but
this feels like a useful small improvement.

The equivalent method in AnyInstanceMethod already works this way and
InstanceMethod does not have an equivalent method.