Use Ubuntu Precise vs Trusty for Ruby v1.8.7 builds
Travis CI is now using Ubuntu Trusty (v14.04) by default. However,
it appears that rvm cannot install Ruby v1.8.7 in this environment.
See the failing jobs in this build [1] and this Github issue [2].

That issue suggests that the fix is to use Ubuntu Precise (v12.04)
for the Ruby v1.8.7 builds which is what this commit attempts to do.

[1]: https://travis-ci.org/freerange/mocha/builds/180746952
[2]: https://github.com/travis-ci/travis-ci/issues/4984