Explicitly instantiate Mockery from Hooks#mocha_setup
Previously the Mockery singleton instance was lazily instantiated the
first time it was needed. This meant that it could be instantiated
outside the scope of a test with no subsequent calls to Mockery.verify,
Mockery.teardown, or Mockery.reset_instance. This had the potential to
leave things in an inconsistent state and resulted in confusing errors
in tests. See #292 for some examples reported by users. This change is a
step on the way to solving that issue.

Note that rather than simply instantiating a single instance of a
Mockery, I've had to store a stack of instances. This is entirely to
cope with the acceptance tests which result in nested calls to
Mockery.setup and Mockery.teardown:

1. Outer acceptance test:              Mockery.setup
2. Inner test in run_as_test(s) block: Mockery.setup
3. Inner test in run_as_test(s) block: Mockery.teardown
4. Outer acceptance test:              Mockery.teardown

I've also had to copy the logger instance from the outer Mockery
instance to the inner one in order that the outer acceptance tests have
access to the same logger as used by the inner one.

Obviously the above complication is a bit undesirable, given that it's
only needed to support the acceptance tests and should never be required
by users. However, I tried various alternatives and I felt this one was
the best option.

Note also that this indirectly addresses #293.