Raise NotInitializedError if Mocha used outside context of a test
Hopefully this addresses #292.

If no Mockery has been instantiated when a call is made to any method
which would otherwise change the state of the Mockery instance then
raise a NotInitializedError.

This is achieved by returning an instance of Mockery::Null if no Mockery
has been instantiated. Note that the Mockery::Null class inherits from
the Mockery class, but overrides all methods which result in a change
of state.