Use add_dependency when add_development_dependency not available
Looking back through the gemspec history it looks as if this was the
intention of these conditional branches. Unfortunately I don't think we
have any tests or CI builds to check these conditional branches, so I'm
just hoping I haven't broken anything. My justification for this
approach is that these conditional branches relate to very old versions
of Rubygems.

Note that I have confirmed that the `add_development_dependency` method
was introduced in Rubygems v1.2.0. [1]

[1]:
https://github.com/rubygems/rubygems/commit/efd062d90585a95f59d311ca9d676a94775be120