Remove setting of gemspec version
The current specification_version is 4. The only change from version 3
to version 4 appears to be that version 4 allows a `metadata` attribute
to be set on the specification [1]. What's more the
specification_version has not been incremented since 2011. Thus I can't
see any advantage to setting the specification_version to 3 rather than
leaving it as the default value which is currently 4. It seems safe to
assume that if the current version is ever incremente, the specification
ought to be backwardly compatible and if not, we'll cross that bridge
when we come to it.

Note that I have removed a conditional branch in this commit. As far as
I can see this branch is identical to the branch where the Rubygems
version is < 1.2.0 which is when the add_development_dependency method
was introduced. I can't see a reason to have a seperate check on the
existence of the specification_version method which was introduced in
Rubygems v0.8.0 [2].

[1]:
https://github.com/rubygems/rubygems/commit/be483b99d8a53bce3e6faf9e6f3c3a5df452974e
[2]:
https://github.com/rubygems/rubygems/commit/1f311b2d5f028d7ec25a625739c4e263657e04cc