Fix Security/YAMLLoad violations using auto-correct
I disabled the cop for the YamlEquivalent matcher, because it's possible
that classes not allowed by safe_load might be passed to the matcher. It
feels safe to do this because it should only ever be used in test code
and not production code.

I'm happy to accept the change suggested by auto-correct in the
build-matrix script, because the Travis config only contains simple
types allowed by safe_load.