Allow long comment lines and regenerate rubocop todo list
Many of the longest lines are comment lines for YARD documentation. I
think splitting the lines might change the formatting of the
documentation, so I'm happy to allow them by setting the IgnoredPatterns
option on the Metrics/LineLength cop.

Note that this reduces the offense count from 735 to 501 and the maximum
line length needed from 545 to 180. While the latter is still too big,
this is a definite improvement.