Set rubocop TargetRubyVersion to earlist possible (v2.2)
Ideally we'd set this to match the `required_ruby_version` in the
gemspec, i.e. v1.8.7. However, we might as well set to to the earliest
possible version in order to avoid triggering cops only intended to
operate in more recent versions of Ruby.

This means it's safe to no longer disable a couple of such cops in
`.rubocop.yml`. I've confirmed this is OK by regenerating the todo list
after making the change and seeing no changes to the todo list.