Remove dependency on metaclass gem
The singleton_class method has been available since Ruby v1.9.2 and so
it seems better to avoid this dependency, especially as it's very simple
to implement it for earlier Ruby versions.

While the singleton_class method is documented on the Object class [1], it's
actually implemented on the Kernel module. See this comment [2] in
`object.c` in the Ruby source and the implementation in the backports
gem [3].

Note that previously the metaclass gem implemented the method using
a double-underscore prefix and suffix which meant that the method was
automatically excluded from the tests involving OBJECT_METHODS in
MockTest. So I've had to explicitly add the singleton_class method to
the list of excluded methods for these tests.

[1]: http://ruby-doc.org/core-1.9.2/Object.html#method-i-singleton_class
[2]: https://github.com/ruby/ruby/blob/c8e9e0b74b7fb2e225af8708426389db88f80683/object.c#L4120
[3]: https://github.com/marcandre/backports/blob/edd1109ae8865d58daf34d256bcf075520f5ffdf/lib/backports/1.9.2/kernel/singleton_class.rb