Refactor: Further DRY up checks for actions (allow/prevent/warn)
on_stubbing_method_unnecessarily differs only in how the method_signature and
backtrace are computed. This change means that those two things are computed
even in case the actions are allowed rather than after the allowed check. A
solution to that would be to turn each check into an object and we could then
have a pipeline of checks we run through.