Fix: Bug in AnyInstanceMethod#restore_original_method
This fixes a bug in AnyInstanceMethod where it would fail to restore the
visibility of a stubbed method if it didn't overwrite the original. This
commit introduces a new acceptance test to highlight the previously
incorrect behaviour.

This also has the advantage of making
AnyInstanceMethod#restore_original_method look more similar to
ClassMethod#restore_original_method.

Co-authored-by: James Mead <james@floehopper.org>