Refactor: rename ClassMethod to StubbedMethod
ClassMethod is/was being used solely as a superclass of InstanceMethod and
AnyInstanceMethod, and it seems odd to have them inherit from ClassMethod. It
seems there're only two kinds of 'stubbed methods' - those that stub an actual/
specific instance (InstanceMethod) and those that stub any instance
(AnyInstanceMethod). The abstract concept represented by both of them is what
I'm calling a StubbedMethod. Not a 100% sure about that name, though.