Refactor: call method_exists? directly on object's singleton_class
Since a method exists (or not) on a class, and an object responds (or not) to a
message, Mockery can directly call method_exists? on the singleton_class of an
object, and continue to call respond_to? on the object itself. There's no need
to ask an object whether a method_exists?, now that we have method_visibility/
method_exists? on a Class (through ClassMethods)