fail yields when the caller does not provide a block
This is more reasonable and realistic than yielding only if a block is given.
A stub that's set up to yield must unconditionally yield, just like it would
unconditionally raise or return when set up to do so. The block_given? check
should be the responsibility of the original method implementation if it so
chooses. A stub set up to yield is presumably set up to simulate the code path
that actually yields.