prepare to pass Invocation to ExpectationList methods
This also simplifies the code and state management in Invocation, as all
required ivar now happens in the constructor and the call method only records
the result of the invocation. This gives us a clear split between the before and
after of an invocation. It doesn't require defaults for Invocation constructor
parameters.