Add display_matching_invocations_on_failure configuration option
So that this configuration option can be set without using the
MOCHA_OPTIONS environment variable which seems overly global in scope.

    Mocha.configure do |c|
      c.display_matching_invocations_on_failure = true
    end

Note that the code using Expectation#verbose? has not yet been released
so there is no need to deprecate anything here. Also I've elected not to
provide old-stlye configuration methods for the new option since these
are now deprecated.