Avoid deprecation warnings when running acceptance tests
I was seeing deprecation warnings like this one:

    Mocha deprecation warning at /Users/jamesmead/Code/freerange/mocha/test/acceptance/bug_21563_test.rb:2:in `require': Use `require 'mocha/test_unit'` or `require 'mocha/minitest'` instead.

I suspect this problem was introduced in this commit [1].

To fix the problem, I've removed the statements requiring 'mocha/setup'
from all the acceptance tests and instead required either
'mocha/minitest' or 'mocha/test_unit' using the same logic in
`test_helper.rb` which decides which test framework to load. This does
mean there's now a tiny bit of duplication across the test helpers, but
I'm not too worried about that for now.

[1]: 36adf880a0b9302fb5e770161c2e962e7e20fe71