Deprecate support for monkey-patched versions of test-unit & minitest
Versions of minitest earlier than v3.3.0 (27 Jul 2012) and versions of
test-unit earlier than v2.5.1 (05 Jul 2012) will not be supported in
future versions of Mocha.

This is long overdue. The current version of minitest is v5.13.0 and the
current version of test-unit is v3.3.4.

Actually removing support will allow considerable simplification of the
codebase.