Rename ClassMethods#method_visibility -> __method_visibility__
c.f. #70

This method is not part of the public API and yet it is included into
every class in API.included.

I think I've run into a clash with this method in a Rails v6.0.0 project
in activesupport/lib/active_support/core_ext/module/redefine_method.rb
which defines Module#method_visibility [1].

This method was introduced in this commit [2] which was added after the
previous release (v0.9.0). I assume this is why we haven't seen the
problem before.

[1]: https://github.com/rails/rails/blob/v6.0.0/activesupport/lib/active_support/core_ext/module/redefine_method.rb#L30-L39
[2]: 6e50d3818607066d2158a8a6d8313d49dd875f97