Ensure ObjectMethods & ClassMethods included when API extended
Cucumber's World method builds an instance of a "world"
object and then *extends* that object with the modules passed into it.
Thus calling World(Mocha::API) does not invoke Mocha::API.included which
in turn means that Mocha::ObjectMethods is not included into Object and
Mocha::ClassMethods is not included into Class.

I think people were getting away with this by accident because they were
usually requiring 'mocha/setup' or similar which includes Mocha::API
into the relevant test case class and thus indirectly invokes
Mocha::API.included.

It seems particularly important to address this given that without this
change the example code in the documentation for Cucumber integration
added to address #430 does not work!

Fixes #433.