Option to reinstate v1.9 Expectation#yields & #multiple_yields behaviour
In v1.9 if Expectation#yields or Expectation#multiple_yields was called
on an expectation, but no block was given when the method was invoked,
the instruction to yield was ignored. Since v1.10, a LocalJumpError is
raised.

The old undocumented behaviour is now available via the new
reinstate_undocumented_behaviour_from_v1_9 configuration option, albeit
with a deprecation warning.