Improvements to docs for StateMachine#is & #is_not
Although strictly speaking StateMachine#is is not really "overloaded",
I've used the YARD overload tag to make it clearer that the method can
be used in two different contexts, i.e. with the result being passed to
Expectation#when and used as if it were an instance of StatePredicate or
passed to Expectation#then and used as an instance of State.