Only inspect objects when necessary
This regression was introduced in v1.10.0 in this commit [1] and is
causing the test failures described in this comment [2].

Prior to the commit mentioned above, the method signature was only
generated at the point it was needed, i.e. after the guard condition in
`Mockery#check`. Since that commit the method signature has *always*
been generated.

See #457 for more details.

I'm not entirely happy with this solution, i.e. passing a Proc into
Mockery#check, but it will do for now.

[1]: cd47bbf8c17b1b945d10b3899fb0a2bce801154c
[2]: https://github.com/theforeman/hammer-cli-foreman/pull/460#issuecomment-561686926