extract stub_owner to override for class methods
We were using stubbee as the class whose methods would be stubbed, and
then stubbing stubbee.any_instance. However, for class method tests, we
would want an object (the class) that can be stubbed by calling .stubs
on it directly. We call that object the stub_owner, which in case of
AnyInstance tests will be stubbee.any_instance (or, after the rename
here, stubbed_instance.any_instance