pull up test_should_allow_stubbing_method_responded_to
from AllowStubbingNonExistent{AnyInstance,Class,Instance}{,Superclass}MethodTest
after a rename to make it more generic. This also adds the respond_to? check to
superclass tests. It wasn't clear why we should check for existing methods on a
class and its superclass, but skip the superclass check for a method to which we
the object responds