method_missing won't throw if ExpectationList empty
This is my attempt to fix https://github.com/freerange/mocha/issues/461
Although _mocha_ does not claim to be thread-safe; this change slowly
makes it more resilient to race conditions through use of mocha.

A method on an object/class may be stubbed & invoked prior to the
expectation being added to the list. In such a scenario, the
method_missing in `mock.rb` would raise an exception for unknown
invocation.

Smply add a if-guard that checks that the expectation list is not empty.

fixes #461