stubbee -> stubba_object, mock_owner -> stubbee
stubbee is the thing that's stubbed - more specifically, the object on
which .stubs or .mocks is called. Therefore, it's also the object that
holds a reference to mocha. That's why it was called mock_owner earlier.

stubba_object is the object through which we get hold of the
stubba_class on which stubbed methods reside. In case of AnyInstance,
that class is the stubba_object itself, while in case of Instance, it's
the singleton class of the stubba_object.